backend/requirements.txt
# FastAPI and Server
fastapi>=0.120.0,<0.121.0
uvicorn[standard]>=0.30.0,<0.39.0

# Database
prisma>=0.15.0,<0.16.0
pymongo>=4.15.0,<5.0.0

# Authentication & Security
python-jose[cryptography]>=3.5.0,<4.0.0
bcrypt>=4.0.0,<6.0.0

# Data Validation
pydantic>=2.0.0,<3.0.0
pydantic[email]>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0

